requirements.txt
fastapi
httpx
openai-agents
pydantic
pytest
pytest-asyncio
python-dotenv
uvicorn[standard]
